net: shell: Add arp command
Move "arp" command to a separate file.
