+++ subsys/net/lib/shell/CMakeLists.txt
@@ -7,6 +7,7 @@ zephyr_library_include_directories(. ${ZEPHYR_BASE}/subsys/net/ip)
 zephyr_library_link_libraries_ifdef(CONFIG_MBEDTLS mbedTLS)
 
 zephyr_library_sources(allocs.c)
+zephyr_library_sources(arp.c)
 
 zephyr_library_include_directories(
   ${ZEPHYR_BASE}/kernel/include
