+++ subsys/net/lib/shell/net_shell.c
@@ -47,10 +47,6 @@ LOG_MODULE_REGISTER(net_shell, LOG_LEVEL_DBG);
 
 #include "ipv6.h"
 
-#if defined(CONFIG_NET_ARP)
-#include "ethernet/arp.h"
-#endif
-
 #if defined(CONFIG_NET_L2_ETHERNET)
 #include <zephyr/net/ethernet.h>
 #endif
@@ -1626,79 +1622,6 @@ static void ipv6_frag_cb(struct net_ipv6_reassembly *reass,
 	(*count)++;
 }
 #endif /* CONFIG_NET_IPV6_FRAGMENT */
-
-#if defined(CONFIG_NET_ARP) && defined(CONFIG_NET_NATIVE)
-static void arp_cb(struct arp_entry *entry, void *user_data)
-{
-	struct net_shell_user_data *data = user_data;
-	const struct shell *sh = data->sh;
-	int *count = data->user_data;
-
-	if (*count == 0) {
-		PR("     Interface  Link              Address\n");
-	}
-
-	PR("[%2d] %d          %s %s\n", *count,
-	   net_if_get_by_iface(entry->iface),
-	   net_sprint_ll_addr(entry->eth.addr, sizeof(struct net_eth_addr)),
-	   net_sprint_ipv4_addr(&entry->ip));
-
-	(*count)++;
-}
-#endif /* CONFIG_NET_ARP */
-
-#if !defined(CONFIG_NET_ARP)
-static void print_arp_error(const struct shell *sh)
-{
-	PR_INFO("Set %s to enable %s support.\n",
-		"CONFIG_NET_NATIVE, CONFIG_NET_ARP, CONFIG_NET_IPV4 and"
-		" CONFIG_NET_L2_ETHERNET", "ARP");
-}
-#endif
-
-static int cmd_net_arp(const struct shell *sh, size_t argc, char *argv[])
-{
-#if defined(CONFIG_NET_ARP)
-	struct net_shell_user_data user_data;
-	int arg = 1;
-#endif
-
-	ARG_UNUSED(argc);
-
-#if defined(CONFIG_NET_ARP)
-	if (!argv[arg]) {
-		/* ARP cache content */
-		int count = 0;
-
-		user_data.sh = sh;
-		user_data.user_data = &count;
-
-		if (net_arp_foreach(arp_cb, &user_data) == 0) {
-			PR("ARP cache is empty.\n");
-		}
-	}
-#else
-	print_arp_error(sh);
-#endif
-
-	return 0;
-}
-
-static int cmd_net_arp_flush(const struct shell *sh, size_t argc,
-			     char *argv[])
-{
-	ARG_UNUSED(argc);
-	ARG_UNUSED(argv);
-
-#if defined(CONFIG_NET_ARP)
-	PR("Flushing ARP cache.\n");
-	net_arp_clear_cache(NULL);
-#else
-	print_arp_error(sh);
-#endif
-
-	return 0;
-}
 
 #if defined(CONFIG_NET_CAPTURE)
 static const struct device *capture_dev;
@@ -6187,12 +6110,6 @@ static int cmd_net_websocket(const struct shell *sh, size_t argc,
 	return 0;
 }
 
-SHELL_STATIC_SUBCMD_SET_CREATE(net_cmd_arp,
-	SHELL_CMD(flush, NULL, "Remove all entries from ARP cache.",
-		  cmd_net_arp_flush),
-	SHELL_SUBCMD_SET_END
-);
-
 SHELL_STATIC_SUBCMD_SET_CREATE(net_cmd_capture,
 	SHELL_CMD(setup, NULL, "Setup network packet capture.\n"
 		  "'net capture setup <remote-ip-addr> <local-addr> <peer-addr>'\n"
@@ -6560,8 +6477,6 @@ SHELL_STATIC_SUBCMD_SET_CREATE(net_cmd_udp,
 );
 
 SHELL_STATIC_SUBCMD_SET_CREATE(net_commands,
-	SHELL_CMD(arp, &net_cmd_arp, "Print information about IPv4 ARP cache.",
-		  cmd_net_arp),
 	SHELL_CMD(capture, &net_cmd_capture,
 		  "Configure network packet capture.", cmd_net_capture),
 	SHELL_CMD(conn, NULL, "Print information about network connections.",
